glue: add interface to destroy the stack
TLDK has its destructor function glue_uninit() which will kill
all tcp streams and send rst to the peer, making the whole stack
exit gracefully. However, when using TLDK as a plugin stack in
gVisor, this destructor would never be called when the program
exits, so we adds a helper to invoke this logic to send rst to the
peer.

Signed-off-by: Tianyu Zhou <albert.zty@antgroup.com>